게시글 본인만 수정 및 삭제 가능하게 하기

+++ api_project/src/main/java/com/kch/api_project/service/PostService.java
@@ -10,6 +10,7 @@ import com.kch.api_project.excepstions.ResourceNotFoundException;
 import com.kch.api_project.global.SecurityConfig;
 import com.kch.api_project.repository.PostRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.context.ApplicationContextException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import com.kch.api_project.repository.AuthRepository;
@@ -55,6 +56,7 @@ public class PostService {
         PostDetailDTO result = PostDetailDTO.builder()
                 .title(tp.getTitle())
                 .body(tp.getBody())
+
                 .created_at(tp.getCreated_at())
                 .updated_at(tp.getUpdated_at())
                 .build();
@@ -71,6 +73,7 @@ public class PostService {
                     PostListDTO.builder()
                             .title(tp.getTitle())
                             .id(tp.getId())
+                            .username(tp.getUser().getUsername())
                             .created_at(tp.getCreated_at())
                             .updated_at(tp.getUpdated_at())
                             .build()
@@ -81,10 +84,31 @@ public class PostService {
 
 
     public void deleteTestPost(long id){
+        // 1. 지금 로그인된 사용자의 이름을 가져온다.
+        String requestUserName = SecurityContextHolder.getContext().getAuthentication().getName();
+        // 2. 삭제하고자 하는 게시글 정보를 가져온다.
+        Post post = postRepository.findById(id)
+                        .orElseThrow(() -> new ResourceNotFoundException(("게시글을 찾을 수 없습니다.")));
+        // 3. 삭제하고자 하는 게시글의 작성자의 이름이, 지금 로그인된 사용자의 이름과 같은지 확인
+        if(!post.getUser().getUsername().equals(requestUserName)){
+            throw new ApplicationContextException("작성자만 게시글을 삭제할 수 있습니다.");
+        }
+
         postRepository.deleteById(id);
     }
 
-    public void patachTestPost(int id, PatachTestPostDTO dto){
+
+
+    public void patachTestPost(Long id, PatachTestPostDTO dto){
+        String requestUserName = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(("게시글을 찾을 수 없습니다.")));
+
+        if(!post.getUser().getUsername().equals(requestUserName)){
+            throw new ApplicationContextException("작성자만 게시글을 수정할 수 있습니다.");
+        }
+
         //수정
         Optional<Post> otp = postRepository.findById((long) id);
         if(otp.isEmpty()) {
